Rename Gumbo class to Main class that executes the main logic of the chatbot
